feat: 중간 커밋

+++ src/main/java/kr/hs/dgsw/SOPO_server_v2/domain/board/service/BoardService.java
@@ -5,7 +5,11 @@ import kr.hs.dgsw.SOPO_server_v2.domain.board.dto.BoardLoadRes;
 import kr.hs.dgsw.SOPO_server_v2.domain.board.dto.BoardUpdateReq;
 import kr.hs.dgsw.SOPO_server_v2.domain.board.entity.BoardEntity;
 import kr.hs.dgsw.SOPO_server_v2.domain.board.repository.BoardRepository;
+import kr.hs.dgsw.SOPO_server_v2.domain.member.entity.MemberEntity;
+import kr.hs.dgsw.SOPO_server_v2.domain.member.enums.MemberCategory;
 import kr.hs.dgsw.SOPO_server_v2.global.error.custom.board.BoardNotFound;
+import kr.hs.dgsw.SOPO_server_v2.global.error.custom.member.UserNotCoincide;
+import kr.hs.dgsw.SOPO_server_v2.global.infra.security.GetCurrentMember;
 import kr.hs.dgsw.SOPO_server_v2.global.response.ResponseData;
 import kr.hs.dgsw.SOPO_server_v2.global.response.Response;
 import lombok.RequiredArgsConstructor;
@@ -18,7 +22,9 @@ import java.util.List;
 @Transactional
 @RequiredArgsConstructor
 public class BoardService {
+
     private final BoardRepository boardRepository;
+    private final GetCurrentMember getCurrentMember;
 
     // 게시글 전체 조회
     public ResponseData<List<BoardLoadRes>> getBoards() {
@@ -32,27 +38,31 @@ public class BoardService {
 
     // 빈 게시글 생성
     public ResponseData<Long> createBoard() {
-        // 토큰 정보 가져오기
+        MemberEntity curMember = getCurrentMember.current();
 
         BoardEntity board = BoardEntity.builder()
                 .boardTitle(null)
                 .boardContent(null)
                 .file(null)
-                // .member()
+                .member(curMember)
                 .build();
+
         return ResponseData.of(HttpStatus.OK, "빈 게시물 생성 완료", board.getBoardId());
     }
 
     // 게시글 업데이트
     public Response loadBoard(Long boardId, BoardUpdateReq updateReq) {
-        // 토큰 정보 가져오기
+        MemberEntity curMember = getCurrentMember.current();
 
         BoardEntity board = boardRepository.findById(boardId)
                 .orElseThrow(() -> BoardNotFound.EXCEPTION);
-        // 만약 현재 로그인 유저와 Load 하려는 유저가 다르다면
+
+        // 만약 현재 로그인 유저와 Load 하려는 유저가 다르고, admin 아니라면
+        if (!board.getMember().getMemberId().equals(curMember.getMemberId()) && curMember.getMemberCategory() == MemberCategory.USER) {
+            throw UserNotCoincide.EXCEPTION;
+        }
 
         board.update(updateReq);
-        BoardLoadRes boardLoadRes = BoardLoadRes.of(board);
         return Response.of(HttpStatus.OK, "게시물 업데이트 완료");
     }
 
@@ -67,10 +77,15 @@ public class BoardService {
     // 게시글 삭제
     @Transactional
     public Response deleteBoard(Long boardId) {
-        // 토큰 정보 받아오기
+        MemberEntity curMember = getCurrentMember.current();
+
         BoardEntity board = boardRepository.findById(boardId)
                 .orElseThrow(() -> BoardNotFound.EXCEPTION);
-        // 만약 만든 사람과 삭제하려는 사람이 일치하지 않고 admin이 아니라면.. error
+
+        // 만약 만든 사람과 삭제하려는 사람이 일치하지 않고 admin 아니라면.. error
+        if (!board.getMember().getMemberId().equals(curMember.getMemberId()) && curMember.getMemberCategory() == MemberCategory.USER) {
+            throw UserNotCoincide.EXCEPTION;
+        }
 
         boardRepository.deleteById(boardId);
         return Response.of(HttpStatus.OK, "게시물 삭제 완료");
